+++ src/math/Vec4.cpp
@@ -7,7 +7,7 @@
  * @param z - represents z-coordinate
  * @param w - represents w-coordinate
  */
-Vec4::Vec4(int x, int y, int z, int w)
+Vec4::Vec4(float x, float y, float z, float w)
 {
     components[0] = x;
     components[1] = y;
@@ -25,7 +25,7 @@ Vec4::Vec4()
 /**
  * Getter for x component
  */
-int Vec4::getX()
+float Vec4::getX()
 {
     return components[0];
 }
@@ -34,7 +34,7 @@ int Vec4::getX()
  * Getter for y component
  */
 
-int Vec4::getY()
+float Vec4::getY()
 {
     return components[1];
 }
@@ -42,7 +42,7 @@ int Vec4::getY()
 /**
  * Getter for z component
  */
-int Vec4::getZ()
+float Vec4::getZ()
 {
     return components[2];
 }
@@ -50,7 +50,7 @@ int Vec4::getZ()
 /**
  * Getter for w component
  */
-int Vec4::getW()
+float Vec4::getW()
 {
     return components[3];
 }
